% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTree.R
\name{cell.ordering.table}
\alias{cell.ordering.table}
\title{Ranking of cells according to backbone tree structure}
\usage{
cell.ordering.table(b.tree, write.to.tex.file = NULL)
}
\arguments{
\item{b.tree}{An \link[igraph]{igraph} backbone tree, as returned by \code{\link{compute.backbone.tree}}.}

\item{write.to.tex.file}{Boolean (optional). If not \code{NULL}, outputs LaTeX version of table to file \code{write.to.tex.file}.}
}
\value{
List of all cells, ranked by position in backbone tree, along with topic information.
}
\description{
Produces a table of input cells ranked by their position in the backbone tree.
}
\examples{
# Load pre-computed LDA model for skeletal myoblast RNA-Seq data from HSMMSingleCell package:
data(HSMM_lda_model)

# Recover sampling time (in days) for each cell:
library(HSMMSingleCell)
data(HSMM_sample_sheet)
days.factor = HSMM_sample_sheet$Hours
days = as.numeric(levels(days.factor))[days.factor]

# Compute near-optimal backbone tree:
b.tree = compute.backbone.tree(HSMM_lda_model, days)
# Load pre-computed LDA model for skeletal myoblast RNA-Seq data from HSMMSingleCell package:
temp.output = tempfile()
cell.ordering.table(b.tree, write.to.tex.file = temp.output)
}

