% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTree.R
\docType{package}
\name{cellTree-package}
\alias{cellTree}
\alias{cellTree-package}
\title{Inference and visualisation of Single-Cell RNA-seq Data data as a hierarchical tree structure}
\description{
This packages computes a Latent Dirichlet Allocation (LDA) model of single-cell RNA-seq data and build a compact tree modelling the relationship between individual cells over time or space.
}
\details{
A typical use-case will require you to run \link{compute.lda} on your expression data, to fit an LDA model, followed by \link{compute.backbone.tree} to generate a tree structure from the LDA model.

Plotting functions \link{ct.plot.grouping} and \link{ct.plot.topics} will then help you plot the tree structure with the desired annotations, while function \link{cell.ordering.table} will output an ordered table of cells, ranked by their position in the tree.

To get further information on each topic, you can run Gene Ontology enrichment tests, using \link{compute.go.enrichment}, plot the result tables as a graph using \link{ct.plot.go.dag} or render it with LaTeX, using \link{go.results.to.latex}.
}

