% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTree.R
\name{compute.go.enrichment}
\alias{compute.go.enrichment}
\title{Gene Ontology enrichment analysis}
\usage{
compute.go.enrichment(lda.results, go.db, ontology.type = "BP",
  reformat.gene.names = FALSE, bonferroni.correct = TRUE,
  p.val.threshold = if (bonferroni.correct) 0.05 else 0.01,
  go.score.class = "weight01Score", dag.file.prefix = FALSE)
}
\arguments{
\item{lda.results}{A fitted LDA model, as returned by \code{\link{compute.lda}}}

\item{go.db}{String. Genome-wide annotation with GO mapping for the appropriate organism (e.g. \pkg{org.Mm.eg.db} or \pkg{org.Hs.eg.db}).}

\item{ontology.type}{(optional). ``BP'' for Biological Process, ``MF'' for Molecular Function, and ``CC'' for Cellular Component.}

\item{reformat.gene.names}{Boolean. If set to \code{TRUE}, converts all gene names to capitalised lowercase.}

\item{bonferroni.correct}{Boolean. Unless set to \code{FALSE}, adjust statistical testing p-value threshold for multiple testing.}

\item{p.val.threshold}{Numeric (optional). P-value significance threshold.}

\item{go.score.class}{String (optional). Name of the scoring method to use for the Kolmogorov-Smirnov test (e.g. ``weigth01Score'' or ``elimScore''). See \pkg{topGO} documentation for a complete list of scoring methods.}

\item{dag.file.prefix}{String or \code{FALSE}. If not set to \code{FALSE}, plots individual subgraphs of significant terms for each topic using the string as filename prefix.}
}
\value{
Returns a named list object with ranked tables of significantly enriched GO terms for each topic (`all'), terms that only appear in each topic (`unique') and terms that appear in less than half of the other topics (`rare'). In addition the list object contains an \link[igraph]{igraph} object with the full GO DAG, annotated with each term's p-value and the significance threshold adjusted for multiple testing (Bonferroni method).
}
\description{
Computes enrichment scores for Gene Ontology terms associated with genes in each topic.
}
\examples{
# Load pre-computed LDA model for skeletal myoblast RNA-Seq data from HSMMSingleCell package:
data(HSMM_lda_model)

\donttest{
# Load GO mapping database for 'homo sapiens':
library(org.Hs.eg.db)
# Compute Cellular Component GO enrichment sets for each topic:
go.results = compute.go.enrichment(HSMM_lda_model, org.Hs.eg.db, ontology.type="CC", bonferroni.correct=TRUE, p.val.threshold=0.01)

# Print table of terms that are only significantly enriched in each topic: 
print(go.results$unique)
}
}

