% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTree.R
\name{ct.plot.go.dag}
\alias{ct.plot.go.dag}
\title{Gene Ontology enrichment sets plotting}
\usage{
ct.plot.go.dag(go.results, up.generations = 2, only.topics = NULL,
  file.output = NULL, p.val.threshold = go.results$adjusted.p.threshold,
  only.unique = FALSE, topic.colors = rainbow(length(go.results$results)))
}
\arguments{
\item{go.results}{GO Enrichment result list object, such as returned by \code{\link{compute.go.enrichment}}.}

\item{up.generations}{Integer (optional). Number of generations above significant nodes to include in the subgraph.}

\item{only.topics}{Integer vector (optional). If not \code{NULL}, vector of topics that should be included in the plot (otherwise all topic enrichment sets are used).}

\item{file.output}{String (optional). If not \code{NULL}, pathname of file to write the plot to.}

\item{p.val.threshold}{Numeric (optional). P-value treshold to use to select which terms should be plotted.}

\item{only.unique}{Only display terms that are only significant for one of the topics.}

\item{topic.colors}{RGB colour vector (optional). Colors to use for each topic.}
}
\value{
An \link[igraph]{igraph} object with the annotated GO DAG.
}
\description{
Plots DAG of significantly enriched terms for all topics, along with ancestor nodes.
}
\examples{
# Load pre-computed LDA model for skeletal myoblast RNA-Seq data from HSMMSingleCell package:
data(HSMM_lda_model)

# Load GO mapping database for 'homo sapiens':
library(org.Hs.eg.db)

\donttest{
# Compute GO enrichment sets for each topic:
go.results = compute.go.enrichment(HSMM_lda_model, org.Hs.eg.db, bonferroni.correct=TRUE)

go.dag.subtree = ct.plot.go.dag(go.results, up.generations = 2)

}
}

