% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTree.R
\name{ct.plot.heatmap}
\alias{ct.plot.heatmap}
\title{Gene Expression Heatmap}
\usage{
ct.plot.heatmap(data, b.tree, log.scale = TRUE, sd.filter = 0.7,
  reorder.genes = TRUE)
}
\arguments{
\item{data}{A matrix of (non-negative) RNA-seq expression levels where each row is a gene and each column is the cell sequenced.}

\item{b.tree}{igraph object returned by \code{\link{compute.backbone.tree}}.}

\item{log.scale}{Boolean (optional). Whether the data should be log-scaled.}

\item{sd.filter}{Numeric or \code{FALSE} (optional). Standard-deviation threshold below which genes should be removed from the data (no filtering if set to \code{FALSE}).}

\item{reorder.genes}{Boolean (optional). Whether the gene rows should be reordered using a dendrogram of their mean value.}
}
\value{
\code{data} object reordered according to the backbone tree, such as used to plot the heatmap.
}
\description{
Plots a heatmap of gene expression, with cells ordered according to the structure computed by \code{\link{compute.backbone.tree}}.
}
\examples{
# Load pre-computed LDA model for skeletal myoblast RNA-Seq data from HSMMSingleCell package:
data(HSMM_lda_model)

# Recover sampling time (in days) for each cell:
library(HSMMSingleCell)
data(HSMM_sample_sheet)
days.factor = HSMM_sample_sheet$Hours
days = as.numeric(levels(days.factor))[days.factor]

# Compute near-optimal backbone tree:
b.tree = compute.backbone.tree(HSMM_lda_model, days)
# Plot heatmap:
data(HSMM_expr_matrix)
ct.plot.heatmap(HSMM_expr_matrix[1:2000,], b.tree, reorder.genes=FALSE)
}

