% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTree.R
\name{get.cell.dists}
\alias{get.cell.dists}
\title{Cell Pairwise-Distance Matrix}
\usage{
get.cell.dists(lda.results)
}
\arguments{
\item{lda.results}{A fitted LDA model, as returned by \code{\link{compute.lda}}}
}
\value{
A square matrix of pairwise distance between cells in the input model.
}
\description{
Computes the pairwise distance between cells, based on the topic histograms form a fitted LDA model.
}
\details{
Distances between histograms are computed using the Chi-square distance
}
\examples{
# Load pre-computed LDA model for skeletal myoblast RNA-Seq data from HSMMSingleCell package:
data(HSMM_lda_model)

# Compute cell pairwise distances:
b.tree = get.cell.dists(HSMM_lda_model)
}

